+++ be/src/Project/Services/Inventory/Inventory.Api/v1/RequestModels/OutletRequestModel.cs
@@ -11,9 +11,9 @@ namespace Inventory.Api.v1.RequestModels
         public string ProductSku { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
         public string Ean { get; set; }
-        public int AvailableQuantity { get; set; }
+        public double AvailableQuantity { get; set; }
         public Guid? OrganisationId { get; set; }
     }
 }
