+++ be/src/Project/Web/Seller/Seller.Web/Areas/Inventory/DomainModels/InventoryItem.cs
@@ -11,10 +11,10 @@ namespace Seller.Web.Areas.Inventory.DomainModels
         public string ProductName { get; set; }
         public string Sku { get; set; }
         public Guid SellerId { get; set; }
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
         public string Ean { get; set; }
         public int? RestockableInDays { get; set; }
-        public int AvailableQuantity { get; set; }
+        public double AvailableQuantity { get; set; }
         public DateTime? ExpectedDelivery { get; set; }
         public DateTime LastModifiedDate { get; set; }
         public DateTime CreatedDate { get; set; }
